+++ jedrock-network/src/main/java/com/jedrock/network/JedrockConnection.java
@@ -127,6 +127,16 @@ public class JedrockConnection implements Connection, PlayerConnection {
     }
 
     @Override
+    public void setSidebar(String title, String[] lines) {
+        protocolHandler.setSidebar(this, title, lines);
+    }
+
+    @Override
+    public void clearSidebar() {
+        protocolHandler.clearSidebar(this);
+    }
+
+    @Override
     public void showHeldItem(long entityId, int state) {
         protocolHandler.showHeldItem(this, entityId, state);
     }